+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Take care that photos to retrieve count is not greater than total photos count (at AbstractPicker level)
 * On FilesystemUploader, check that target directory exists and is empty before scanning
 * Add code coverage when running `make test`
+* Allow to specify an order for retrieved photos
 
 # 0.3.1
 * Allow to use a custom target directory in Dropbox